LPS-152805 Keeps logic only in display context class
